+++ actions/get_version.py
@@ -15,8 +15,5 @@ class GetVersion(action.TerraformBaseAction):
 
         self.terraform.terraform_bin_path = terraform_exec
         return_code, stdout, stderr = self.terraform.version()
-        output = stdout + "\n" + stderr
-        if return_code == 0:
-            return (True, output)
-        else:
-            return (False, output)
+
+        return self.check_result(return_code, stdout, stderr)