+++ tests/terraform_base_action_test_case.py
@@ -1,0 +1,5 @@
+from st2tests.base import BaseActionTestCase
+
+
+class TerraformBaseActionTestCase(BaseActionTestCase):
+    __test__ = False